fix(frontend): resolve cross-cutting lint and only-export-components
- ScreenshotViewer: drop always-1 carousel index + ref-during-render reset
- HomePage: pick completion message via a stable mounted seed (pure render)
- Replace raw bg-gray-950 with semantic bg-background token
- Move user-analytics number formatters to a dedicated module so the
  components file only exports components (only-export-components)
